moving button.tsx to outside folder
