+++ components/appbar/index.tsx
@@ -11,7 +11,7 @@ import ListItemText from '@mui/material/ListItemText';
 import MenuIcon from '@mui/icons-material/Menu';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-import TealButton from '../button/index';
+import TealButton from '../button';
 import Button from '@mui/material/Button';
 
 interface Props {
